Refactoring new files that are added after pulling from master
