+++ docs/source/flow.envs.bay_bridge.rst
@@ -4,7 +4,7 @@ flow.envs.bay\_bridge package
 Submodules
 ----------
 
-flow.envs.bay\_bridge.base module
+flow.envs.bay\_bridge module
 ---------------------------------
 
 .. automodule:: flow.envs.bay_bridge
