+++ examples/aimsun/small_template.py
@@ -4,7 +4,7 @@ from flow.core.experiment import Experiment
 from flow.core.params import AimsunParams, EnvParams, NetParams
 from flow.core.params import VehicleParams
 from flow.envs import TestEnv
-from flow.scenarios.loop import Scenario
+from flow.scenarios.ring import Scenario
 from flow.core.params import InFlows
 import flow.config as config
 import os
