+++ examples/stable_baselines/stabilizing_highway.py
@@ -90,7 +90,7 @@ flow_params = dict(
     exp_tag="stabilizing_open_network_merges",
 
     # name of the flow environment the experiment is running on
-    env_name="WaveAttenuationMergePOEnv",
+    env_name="MergePOEnv",
 
     # name of the scenario class the experiment is running on
     scenario="MergeScenario",
