+++ examples/stable_baselines/stabilizing_the_ring.py
@@ -46,7 +46,7 @@ flow_params = dict(
     env_name="WaveAttenuationPOEnv",
 
     # name of the scenario class the experiment is running on
-    scenario="LoopScenario",
+    scenario="RingScenario",
 
     # simulator that is used by the experiment
     simulator='traci',
