+++ examples/stable_baselines/velocity_bottleneck.py
@@ -108,7 +108,7 @@ flow_params = dict(
     exp_tag="DesiredVelocity",
 
     # name of the flow environment the experiment is running on
-    env_name="DesiredVelocityEnv",
+    env_name="BottleneckDesiredVelocityEnv",
 
     # name of the scenario class the experiment is running on
     scenario="BottleneckScenario",
