+++ flow/envs/multiagent/highway.py
@@ -2,7 +2,7 @@
 import numpy as np
 from gym.spaces.box import Box
 from flow.core.rewards import desired_velocity
-from flow.multiagent_envs.multiagent_env import MultiEnv
+from flow.envs.multiagent.base import MultiEnv
 
 
 ADDITIONAL_ENV_PARAMS = {
