+++ tests/fast_tests/test_examples.py
@@ -4,9 +4,8 @@ import unittest
 import ray
 from ray.tune import run_experiments
 
-from examples.rllib.cooperative_merge import setup_exps as coop_setup
 from examples.rllib.figure_eight import setup_exps as figure_eight_setup
-from examples.rllib.green_wave import setup_exps as green_wave_setup
+from examples.rllib.traffic_light_grid import setup_exps as traffic_light_grid_setup
 from examples.rllib.stabilizing_highway import setup_exps as highway_setup
 from examples.rllib.stabilizing_the_ring import setup_exps as ring_setup
 from examples.rllib.velocity_bottleneck import setup_exps as bottleneck_setup
@@ -24,7 +23,7 @@ from examples.rllib.multiagent_exps.multiagent_highway import setup_exps \
     as multi_highway_setup
 
 from examples.stable_baselines.figure_eight import run_model as run_figure_eight
-from examples.stable_baselines.green_wave import run_model as run_green_wave
+from examples.stable_baselines.traffic_light_grid import run_model as run_traffic_light_grid
 from examples.stable_baselines.stabilizing_highway import run_model as run_stabilizing_highway
 from examples.stable_baselines.stabilizing_the_ring import run_model as run_stabilizing_ring
 from examples.stable_baselines.velocity_bottleneck import run_model as run_velocity_bottleneck
@@ -34,10 +33,9 @@ from examples.sumo.bay_bridge_toll import bay_bridge_toll_example
 from examples.sumo.bottlenecks import bottleneck_example
 from examples.sumo.density_exp import run_bottleneck
 from examples.sumo.figure_eight import figure_eight_example
-from examples.sumo.grid import grid_example
+from examples.sumo.traffic_light_grid import grid_example
 from examples.sumo.highway import highway_example
 from examples.sumo.highway_ramps import highway_ramps_example
-from examples.sumo.loop_merge import loop_merge_example
 from examples.sumo.merge import merge_example
 from examples.sumo.minicity import minicity_example
 from examples.sumo.sugiyama import sugiyama_example
@@ -70,8 +68,8 @@ class TestSumoExamples(unittest.TestCase):
         # run the experiment for a few time steps to ensure it doesn't fail
         exp.run(1, 5)
 
-    def test_grid(self):
-        """Verifies that examples/sumo/grid.py is working."""
+    def test_traffic_light_grid(self):
+        """Verifies that examples/sumo/traffic_light_grid.py is working."""
         # test the example in the absence of inflows
         exp = grid_example(render=False, use_inflows=False)
         exp.run(1, 5)
@@ -159,8 +157,8 @@ class TestStableBaselineExamples(unittest.TestCase):
         This is done by running each experiment in that folder for five time-steps
         and confirming that it completes one rollout with two workers.
     """
-    def test_run_green_wave(self):
-        run_green_wave(num_steps=5)
+    def test_run_traffic_light_grid(self):
+        run_traffic_light_grid(num_steps=5)
 
     def test_run_figure_eight(self):
         run_figure_eight(num_steps=5)
@@ -186,22 +184,18 @@ class TestRllibExamples(unittest.TestCase):
         if not ray.is_initialized():
             ray.init(num_cpus=1)
 
-    def test_coop_merge(self):
-        alg_run, env_name, config = coop_setup()
-        self.run_exp(alg_run, env_name, config)
-
     def test_figure_eight(self):
         alg_run, env_name, config = figure_eight_setup()
         self.run_exp(alg_run, env_name, config)
 
-    def test_green_wave(self):
+    def test_traffic_light_grid(self):
         # test the example in the absence of inflows
-        alg_run, env_name, config = green_wave_setup(use_inflows=False)
+        alg_run, env_name, config = traffic_light_grid_setup(use_inflows=False)
         self.run_exp(alg_run, env_name, config)
 
-    def test_green_wave_inflows(self):
+    def test_traffic_light_grid_inflows(self):
         # test the example in the presence of inflows
-        alg_run, env_name, config = green_wave_setup(use_inflows=True)
+        alg_run, env_name, config = traffic_light_grid_setup(use_inflows=True)
         self.run_exp(alg_run, env_name, config)
 
     def test_stabilizing_highway(self):
